autopep + fix can
