+++ sinks/dashboard/parsers.py
@@ -2,6 +2,8 @@ from series import publisher
 
 _func_map = {}
 # decorator for parse functions to save a massive if chain
+
+
 def register(msg_channels):
     if isinstance(msg_channels, str):
         msg_channels = [msg_channels]
@@ -22,7 +24,7 @@ def daq_parser(msg_data):
 
     for sensor, data in msg_data["data"].items():
         parsed_messages.append((sensor, timestamp, sum(data)/len(data)))
-    
+
     return parsed_messages
 
 
@@ -31,12 +33,15 @@ def parse(msg_channel, msg_payload):
         if msg_channel.startswith(func):
             series_message_pair_list = _func_map[func](msg_payload)
             for series_name, timestamp, parsed_message in series_message_pair_list:
-                publisher.update(series_name,[timestamp, parsed_message])
+                publisher.update(series_name, [timestamp, parsed_message])
+
 
 ######################################## OLD CODE #############################################
 #
 BOARD_NAME_LIST = ["DUMMY", "INJECTOR", "LOGGER", "RADIO", "SENSOR", "VENT", "GPS", "ARMING",
                    "PAPA", "ROCKET_PI", "ROCKET_PI_2", "SENSOR_2", "SENSOR_3"]
+
+
 @register(BOARD_NAME_LIST)
 def can_parser(msg_data):
     timestamp = msg_data["timestamp"]
@@ -46,7 +51,7 @@ def can_parser(msg_data):
 #
 #from series import CanMsgSeries
 #
-#class Parser:
+# class Parser:
 #    """
 #    Turns omnibus messages into series of their data
 #    """
@@ -94,7 +99,7 @@ def can_parser(msg_data):
 #            return None
 #        return Parser.parsers[channel][0].series[name]  # SeriesDefaultDict takes care of the rest
 #
-#class CanDisplayParser(Parser):
+# class CanDisplayParser(Parser):
 
 #
 #    def __init__(self):
@@ -116,4 +121,4 @@ def can_parser(msg_data):
 #        return CanDisplayParser.canSeries.values()
 #
 #
-#CanDisplayParser()
+# CanDisplayParser()
