+++ sinks/dashboard/utils.py
@@ -1,6 +1,7 @@
 from pyqtgraph.Qt import QtWidgets
 
-def prompt_user(widget, property_name, description, prompt_type, items=None, can_add_items = False):
+
+def prompt_user(widget, property_name, description, prompt_type, items=None, can_add_items=False):
     """
     Opens a pop up asking user for input. Returns None if input selection
     is not valid or user cancels
@@ -12,8 +13,9 @@ def prompt_user(widget, property_name, description, prompt_type, items=None, can
         if (items == None):
             return None
 
-        selection, ok = QtWidgets.QInputDialog.getItem(widget, property_name, description, items, 0, can_add_items)    
-        
+        selection, ok = QtWidgets.QInputDialog.getItem(
+            widget, property_name, description, items, 0, can_add_items)
+
     elif prompt_type == "text":
         selection, ok = QInputDialog.getText(widget, property_name, description)
